+++ requirements_checker/__init__.py
@@ -1,1 +1,3 @@
-from .main import *
+from .check_string import *
+from .check_requirements_txt import *
+from .check_version import *